# Conflicts:
#	src/main/java/com/softjourn/vending/entity/Product.java

+++ src/main/java/com/softjourn/vending/dao/ImageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
 
-public interface ImageRepository extends CrudRepository<Image,Integer> {
+public interface ImageRepository extends CrudRepository<Image, Long> {
 
     List<Image> findByProductId(Integer id);
 }
